+++ src/lib.rs
@@ -73,6 +73,7 @@ pub mod bitreader;
 pub mod encoder;
 pub mod header;
 pub mod linepack;
+pub(crate) mod spec_exc;
 pub(crate) mod spec_lsp;
 pub mod spec_tables;
 pub mod tables;